Refactor more infrastructure to model_building and migrate tests
