+++ espei/tests/test_model_building.py
@@ -7,7 +7,7 @@ from collections import OrderedDict
 import sympy
 from pycalphad import variables as v
 
-from espei.parameter_selection.model_building import build_feature_sets, build_candidate_models
+from espei.parameter_selection.model_building import build_feature_sets, build_candidate_models, generate_symmetric_group, sorted_interactions
 
 def test_build_feature_sets_generates_desired_binary_features_for_cp_like():
     """Binary feature sets can be correctly generated for heat capacity-like features"""
@@ -127,3 +127,48 @@ def test_ternary_candidate_models_are_constructed_correctly():
             [V_I*YS, V_J*YS, V_K*YS]
         ])
     ])
+
+def test_symmetric_group_can_be_generated_for_2_sl_mixing_with_symmetry():
+    """A phase with two sublattices that are mixing should generate a cross interaction"""
+    symm_groups = generate_symmetric_group((('AL', 'CO'), ('AL', 'CO')), [[0, 1]])
+    assert symm_groups == [(('AL', 'CO'), ('AL', 'CO'))]
+
+
+def test_symmetric_group_can_be_generated_for_2_sl_endmembers_with_symmetry():
+    """A phase with symmetric sublattices should find a symmetric endmember """
+    symm_groups = generate_symmetric_group(('AL', 'CO'), [[0, 1]])
+    assert symm_groups == [('AL', 'CO'), ('CO', 'AL')]
+
+
+def test_interaction_sorting_is_correct():
+    """High order (order >= 3) interactions should sorted correctly"""
+    # Correct sorting of n-order interactions should sort first by number of
+    # interactions of order n, then n-1, then n-2... to 1
+    unsorted_interactions = [
+        ('AL', ('AL', 'CO', 'CR')),
+        (('AL', 'CO'), ('AL', 'CO', 'CR')),
+        (('AL', 'CO', 'CR'), ('AL', 'CO', 'CR')),
+        (('AL', 'CO', 'CR'), 'AL'),
+        (('AL', 'CO', 'CR'), ('AL', 'CO')),
+        (('AL', 'CO', 'CR'), ('AL', 'CR')),
+        (('AL', 'CO', 'CR'), 'CO'),
+        (('AL', 'CO', 'CR'), ('CO', 'CR')),
+        (('AL', 'CO', 'CR'), 'CR'),
+        (('AL', 'CR'), ('AL', 'CO', 'CR')),
+        ('CO', ('AL', 'CO', 'CR')),
+        (('CO', 'CR'), ('AL', 'CO', 'CR')),
+        ('CR', ('AL', 'CO', 'CR')),
+    ]
+    interactions = sorted_interactions(unsorted_interactions, max_interaction_order=3, symmetry=None)
+
+    # the numbers are the different sort scores. Two of the same sort scores mean
+    # the order doesn't matter
+    assert interactions == [
+        ('AL', ('AL', 'CO', 'CR')),                # (1, 0, 1)
+        (('AL', 'CO', 'CR'), 'AL'),                # (1, 0, 1)
+        (('AL', 'CO', 'CR'), 'CO'),                # (1, 0, 1)
+        (('AL', 'CO', 'CR'), 'CR'),                # (1, 0, 1)
+        ('CO', ('AL', 'CO', 'CR')),                # (1, 0, 1)
+        ('CR', ('AL', 'CO', 'CR')),                # (1, 0, 1)
+    ]
+
