+++ espei/tests/test_parameter_generation.py
@@ -4,7 +4,7 @@ from tinydb import where
 
 from espei.tests.fixtures import datasets_db
 
-from espei.paramselect import generate_parameters, generate_symmetric_group, sorted_interactions
+from espei.paramselect import generate_parameters
 from espei.tests.testing_data import *
 
 import pytest
@@ -294,51 +294,6 @@ def test_sgte_reference_state_naming_is_correct_for_character_element(datasets_d
     assert dbf.symbols['GBCCV'].args[0][0].__str__() == 'GHSERVV'
     assert 'GHSERVV' in dbf.symbols.keys()
     assert 'GHSERAL' in dbf.symbols.keys()
-
-
-def test_symmetric_group_can_be_generated_for_2_sl_mixing_with_symmetry():
-    """A phase with two sublattices that are mixing should generate a cross interaction"""
-    symm_groups = generate_symmetric_group((('AL', 'CO'), ('AL', 'CO')), [[0, 1]])
-    assert symm_groups == [(('AL', 'CO'), ('AL', 'CO'))]
-
-
-def test_symmetric_group_can_be_generated_for_2_sl_endmembers_with_symmetry():
-    """A phase with symmetric sublattices should find a symmetric endmember """
-    symm_groups = generate_symmetric_group(('AL', 'CO'), [[0, 1]])
-    assert symm_groups == [('AL', 'CO'), ('CO', 'AL')]
-
-
-def test_interaction_sorting_is_correct():
-    """High order (order >= 3) interactions should sorted correctly"""
-    # Correct sorting of n-order interactions should sort first by number of
-    # interactions of order n, then n-1, then n-2... to 1
-    unsorted_interactions = [
-        ('AL', ('AL', 'CO', 'CR')),
-        (('AL', 'CO'), ('AL', 'CO', 'CR')),
-        (('AL', 'CO', 'CR'), ('AL', 'CO', 'CR')),
-        (('AL', 'CO', 'CR'), 'AL'),
-        (('AL', 'CO', 'CR'), ('AL', 'CO')),
-        (('AL', 'CO', 'CR'), ('AL', 'CR')),
-        (('AL', 'CO', 'CR'), 'CO'),
-        (('AL', 'CO', 'CR'), ('CO', 'CR')),
-        (('AL', 'CO', 'CR'), 'CR'),
-        (('AL', 'CR'), ('AL', 'CO', 'CR')),
-        ('CO', ('AL', 'CO', 'CR')),
-        (('CO', 'CR'), ('AL', 'CO', 'CR')),
-        ('CR', ('AL', 'CO', 'CR')),
-    ]
-    interactions = sorted_interactions(unsorted_interactions, max_interaction_order=3, symmetry=None)
-
-    # the numbers are the different sort scores. Two of the same sort scores mean
-    # the order doesn't matter
-    assert interactions == [
-        ('AL', ('AL', 'CO', 'CR')),                # (1, 0, 1)
-        (('AL', 'CO', 'CR'), 'AL'),                # (1, 0, 1)
-        (('AL', 'CO', 'CR'), 'CO'),                # (1, 0, 1)
-        (('AL', 'CO', 'CR'), 'CR'),                # (1, 0, 1)
-        ('CO', ('AL', 'CO', 'CR')),                # (1, 0, 1)
-        ('CR', ('AL', 'CO', 'CR')),                # (1, 0, 1)
-    ]
 
 
 def test_symmetric_ternary_parameter_can_be_generated(datasets_db):
